added build-version + fixed some build script stuff

+++ scripts/pack.ps1
@@ -1,12 +1,15 @@
-# Define variables
+# vars
 $outputDir = "./build/"
 $iconPath = "./assets/icons"
 $appName = "BazFlipper"
+$buildVersion = "0.2.0"
 
-Remove-Item -LiteralPath $outputDir -Force -Recurse
+# cleanup old builds
+if (Test-Path "./build" -PathType Container) {
+    Remove-Item -LiteralPath $outputDir -Force -Recurse
+}
 
-# Package for Windows
-electron-packager . $appName --platform=win32 --arch=x64 --out=$outputDir --icon=$iconPath/icon.ico
 
-# Package for Linux
-electron-packager . $appName --platform=linux --arch=x64 --out=$outputDir --icon=$iconPath/icon.png
+# build
+electron-packager . $appName --platform=win32 --arch=x64 --build-version=$buildVersion --out=$outputDir --icon=$iconPath/icon.ico
+electron-packager . $appName --platform=linux --arch=x64 --out=$outputDir --icon=$iconPath/icon.png
